feat(user-portal): introduce routing
