+++ packages/wip-user-portal/package.json
@@ -31,7 +31,9 @@
     "@types/react": "16.7.18",
     "@types/react-dom": "16.0.11",
     "react": "16.8.1",
-    "react-dom": "16.8.1"
+    "react-dom": "16.8.1",
+    "react-router": "4.3.1",
+    "react-router-dom": "4.3.1"
   },
   "devDependencies": {
     "@babel/core": "7.2.2",
@@ -39,6 +41,7 @@
     "@babel/preset-env": "7.3.1",
     "@babel/preset-react": "7.0.0",
     "@types/express": "4.16.1",
+    "@types/react-router-dom": "4.3.1",
     "bundlesize": "0.17.1",
     "nodemon": "1.18.10",
     "webpack": "4.29.3",
