--- packages/wip-user-portal/src/app/App.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import { CashierApp } from './Cashier/CashierApp';
-
-export const App = () => <CashierApp />;
