+++ packages/wip-user-portal/src/app/cashier/cashier-app.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import { Global } from '@emotion/core';
+import { styles } from '@heathmont/sportsbet-global';
+
+import { CashierRoutes } from './routes';
+
+export const CashierApp = () => (
+  <React.Fragment>
+    <Global styles={styles} />
+    <CashierRoutes />
+  </React.Fragment>
+);
