+++ packages/wip-user-portal/src/app/cashier/views/exchange/dumb-components/heading.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { Heading } from '@heathmont/sportsbet-components';
+
+export const CashierHeading = () => (
+  <Heading size="foxtrot" element="h1">
+    Cashier Heading
+  </Heading>
+);
