+++ packages/wip-user-portal/src/client.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { App } from './app/App';
+import { ClientSideApp } from './app/app';
 
 const app = document.getElementById('app');
-ReactDOM.hydrate(<App />, app);
+ReactDOM.hydrate(<ClientSideApp />, app);
